% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cmp_function_dependence}
\alias{cmp_function_dependence}
\title{A comparator ('<=' equivalent) for quicksort.}
\usage{
cmp_function_dependence(f, g)
}
\arguments{
\item{f}{A function.}

\item{g}{Another function.}
}
\value{
Returns TRUE if g does not depend on f.
}
\description{
Compares functions based on dependencies!
}

