% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{db_connector-class}
\alias{db_connector}
\alias{db_connector-class}
\title{A reference class which always yields a valid connection with x$conn.}
\description{
A reference class which always yields a valid connection with x$conn.
}
\section{Fields}{

\describe{
\item{\code{driver}}{A driver from RPostgreSQL.}

\item{\code{credentials}}{A filename for connection info. to the database
(including host and port).}

\item{\code{conn}}{The always-valid connection.}
}}
\section{Methods}{

\describe{
\item{\code{add_connection()}}{Adds a connection to the pool if necessary.}

\item{\code{check_crd(crd = NULL)}}{Checks whether credentials are available and complete.}

\item{\code{clean_connections()}}{Cleans out available connections.}

\item{\code{clear_connections()}}{Runs a simple command to clear pending results on available connections.}

\item{\code{connection_in_use(connection = NULL)}}{Tests whether a connection is in use.}

\item{\code{get_connection()}}{Yields an unused connection from the pool.}

\item{\code{initialize(credentials)}}{Checks for RPostgreSQL, uses credentials to generate a connection and test it.}
}}

