% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dependency_resolver}
\alias{dependency_resolver}
\title{Dependency resolver for function lists.}
\usage{
dependency_resolver(f_list)
}
\arguments{
\item{f_list}{A list of functions to be sorted.}
}
\description{
Returns the function with independent functions coming
first.
}
\examples{
test_list <- list(
	g = function(f, q) { f+q },
	f = function(x) { x^2 },
	h = function() { 10 }
)

sorted_test_list <- dependency_resolver(test_list)
}
\seealso{
cmp_function_dependence, quicksort.list
}

