% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eval_pipe}
\alias{eval_pipe}
\title{Runs expressions from a list in an independently scoped environment.}
\usage{
eval_pipe(data = NULL, pipeline = list())
}
\arguments{
\item{data}{A list of environments (or data frames, or a data frame
which is promoted to an environment.  This creates a scoped
context for evaluation expressions.  This scope \emph{skips}
.GlobalEnv.}

\item{pipeline}{A quoted list of expressions which are evaluted in
scoped context provided by the data.}
}
\value{
An internally generated list of expression results.
}
\description{
Runs expressions from a list in an independently scoped environment.
}
\examples{
data <- replicate(8,new.env())
data[[2]]$z <- pi
pipe=quote(list({a <- 3+3}, {b <- a*2}, {q <-a*b*z} ))
o <- eval_pipe(data,pipe)
}

