% Generated by roxygen2 (4.0.2): do not edit by hand
\name{map}
\alias{map}
\title{Maps names in a list (or data frame) to new names based on a mapping
from a list (map).}
\usage{
map(x, map, one_to_one = TRUE)
}
\arguments{
\item{x}{A list (or data frame) with names to be mapped.}

\item{map}{A named list of character vectors.  In re-mapping list names are the new
target columns and source columns are named in the vectors.}

\item{one_to_one}{A flag which (if TRUE) triggers a warning when
multiple columns might be mapped (only the first is).}
}
\value{
The original list (or data frame) modified.
}
\description{
Maps names in a list (or data frame) to new names based on a mapping
from a list (map).
}
\examples{
n <- 10
widgets <- data.frame(id=1:n, weight_dry=exp(rnorm(n)),
                      weight_wet=exp(rnorm(n))+10,
                      vol=exp(rnorm(n)))
widgets <- map(x=widgets, map=list(volume="vol"))
}

