% Generated by roxygen2 (4.0.2): do not edit by hand
\name{map_unknowns}
\alias{map_unknowns}
\title{Maps special "NA" values to NA's in a vector.}
\usage{
map_unknowns(x, map)
}
\arguments{
\item{x}{A vector with special values.}

\item{map}{A vector of special values to be mapped to NA.}
}
\value{
x, with special values turned to NA.
}
\description{
Maps special "NA" values to NA's in a vector.
}

