% Generated by roxygen2 (4.0.2): do not edit by hand
\name{map_values}
\alias{map_values}
\title{Maps vectorvalues one-to-one.}
\usage{
map_values(x, input, output)
}
\arguments{
\item{x}{A vector to operate on.}

\item{input}{A vector of values to be modified.}

\item{output}{A vector of values to replace input values.}

\item{x,}{with modifications.}
}
\description{
Maps vectorvalues one-to-one.
}

