% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quicksort}
\alias{quicksort}
\alias{quicksort.list}
\title{Sorting with an arbitrarily defined comparator ('<=' by
default).}
\usage{
quicksort(x, cmp = `<=`)
}
\arguments{
\item{x}{The thing to be sorted.}

\item{cmp}{The comparator, '<=' or similar.}
}
\value{
x, but sorted according to cmp.
}
\description{
A quicksort implementation.  It's generic so with the
             right comparator it will do dependency sorting on
             function lists...
}
\details{
Naive, after reading a few web pages about how to do it...
   I just need to sort a short list with a given comparator...
   Based on:

  http://algs4.cs.princeton.edu/23quicksort/
  http://en.wikipedia.org/wiki/Quicksort
  http://rosettacode.org/wiki/Sorting_algorithms/Quicksort
   Thanks internet, that Pascal intro course was a long time ago...
}
\examples{
o <- quicksort(rbinom(n=30, size=15, prob=0.8))
}

