% Generated by roxygen2 (4.0.2): do not edit by hand
\name{safe_read_csv}
\alias{safe_read_csv}
\title{Read a .csv file with minimal modifications as allowed by R,
then modify according to locally sourced files.}
\usage{
safe_read_csv(file, instructions = NULL, drop_uppercase = FALSE)
}
\arguments{
\item{file}{A .csv type file.}

\item{instructions}{A .R file (source-able) which defines column name
mapping ("column_map"), NA mapping ("unknowns"), as well as value
mapping ("value_map", with "input", and "output" vectors.}

\item{drop_uppercase}{Turns column names to all lower-case.}
}
\value{
A data.frame with the .csv file loaded and modified as
specified.
}
\description{
Read a .csv file with minimal modifications as allowed by R,
then modify according to locally sourced files.
}

