% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{standard_string_transformations}
\alias{standard_string_transformations}
\title{Standard strings transformations.}
\format{\preformatted{List of 4
 $ to_lower_case           :function (x)  
 $ drop_leading_whitespace :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 5 28 5 94 35 101 5 5
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x2426380> 
 $ drop_trailing_whitespace:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 6 29 6 95 36 102 6 6
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x2426380> 
 $ whitespace_to_underscore:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 7 29 7 93 36 100 7 7
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x2426380> 
}}
\usage{
standard_string_transformations
}
\description{
Standard strings transformations.
}
\keyword{datasets}

